+++ src/main/java/org/trebol/jpa/services/crud/impl/SalespeopleCrudServiceImpl.java
@@ -24,7 +24,7 @@ import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.trebol.api.models.SalespersonPojo;
+import org.trebol.api.models.PersonPojo;
 import org.trebol.common.exceptions.BadInputException;
 import org.trebol.jpa.entities.Salesperson;
 import org.trebol.jpa.repositories.SalespeopleRepository;
@@ -38,7 +38,7 @@ import java.util.Optional;
 @Transactional
 @Service
 public class SalespeopleCrudServiceImpl
-    extends CrudGenericService<SalespersonPojo, Salesperson>
+    extends CrudGenericService<PersonPojo, Salesperson>
     implements SalespeopleCrudService {
     private final SalespeopleRepository salespeopleRepository;
 
@@ -53,8 +53,8 @@ public class SalespeopleCrudServiceImpl
     }
 
     @Override
-    public Optional<Salesperson> getExisting(SalespersonPojo input) throws BadInputException {
-        String idNumber = input.getPerson().getIdNumber();
+    public Optional<Salesperson> getExisting(PersonPojo input) throws BadInputException {
+        String idNumber = input.getIdNumber();
         if (StringUtils.isBlank(idNumber)) {
             throw new BadInputException("Salesperson does not have an ID card");
         } else {