+++ src/test/java/org/trebol/api/controllers/DataSalespeopleControllerTest.java
@@ -29,7 +29,7 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.trebol.api.DataCrudGenericControllerTest;
-import org.trebol.api.models.SalespersonPojo;
+import org.trebol.api.models.PersonPojo;
 import org.trebol.api.services.PaginationService;
 import org.trebol.jpa.entities.Salesperson;
 import org.trebol.jpa.services.SortSpecParserService;
@@ -48,7 +48,7 @@ import static org.trebol.testing.TestConstants.ANY;
 
 @ExtendWith(MockitoExtension.class)
 class DataSalespeopleControllerTest
-    extends DataCrudGenericControllerTest<SalespersonPojo, Salesperson> {
+    extends DataCrudGenericControllerTest<PersonPojo, Salesperson> {
     @InjectMocks DataSalespeopleController instance;
     @Mock PaginationService paginationServiceMock;
     @Mock SortSpecParserService sortServiceMock;
@@ -77,19 +77,19 @@ class DataSalespeopleControllerTest
 
     @Test
     void creates_salespeople() {
-        assertDoesNotThrow(() -> super.creates_data(SalespersonPojo.builder().build()));
+        assertDoesNotThrow(() -> super.creates_data(PersonPojo.builder().build()));
     }
 
     @Test
     void updates_salespeople_using_predicate_filters_map() {
         assertDoesNotThrow(() -> {
-            SalespersonPojo existingList = SalespersonPojo.builder().build();
-            SalespersonPojo input = SalespersonPojo.builder().build();
+            PersonPojo existingList = PersonPojo.builder().build();
+            PersonPojo input = PersonPojo.builder().build();
             Predicate predicate = new BooleanBuilder();
             when(predicateServiceMock.parseMap(anyMap())).thenReturn(predicate);
             when(crudServiceMock.update(any(), any(Predicate.class))).thenReturn(Optional.of(existingList));
 
-            instance.update(SalespersonPojo.builder().build(), Map.of(ANY, ANY));
+            instance.update(PersonPojo.builder().build(), Map.of(ANY, ANY));
 
             verify(crudServiceMock).update(input, predicate);
         });